Devices: Added CRUD API support

+++ database/migrations/2016_06_29_144854_create_devices_table.php
@@ -19,7 +19,7 @@ class CreateDevicesTable extends Migration
             $table->string('description');
 
             $table->dateTime('purchased_at');
-            $table->double('cost', 7, 2);
+            $table->integer('cost');
 
             $table->text('notes');
